IR Volume control per channel

+++ InputChannelSelector.h
@@ -8,19 +8,19 @@ public:
 	InputChannelSelector(DigitalPotmeter& digitalPotmeter);
 
     int8_t getChannel();
-	void selectChannel(int channel); // 0..7
+	void selectChannel(int8_t channel); // 0..7
 	void selectNextChannel();
 	void selectPreviousChannel();
-	void setDefaultBrightness(int brightness);
+	void setDefaultBrightness(int8_t brightness);
 	void enableDisplay();
 	void disableDisplay();
 
 private:
-	void setBrightness(int brightness);
-	void switchToChannel(int channel);
+	void setBrightness(int8_t brightness);
+	void switchToChannel(int8_t channel);
 
 private:
 	DigitalPotmeter& m_digitalPotmeter;
 	int m_brightness;
-	int m_currentChannel;
+	int8_t m_currentChannel;
 };
